requirements.txt
fastapi==0.61.1
psycopg2-binary==2.8.6
uvicorn==0.12.1
requests==2.24.0
beautifulsoup4==4.9.2
pyexcelerate==0.9.0
lxml==4.5.2
pydantic==1.6.1
python-multipart==0.0.5
sentry_sdk==0.17.6
asyncpg==0.21.0
tortoise-orm==0.16.16
aerich==0.2.5
aiohttp==3.6.2
celery==5.0.0
redis
nest_asyncio==1.4.1
aiofiles==0.5.0
